Support for abbreviated months with trailing periods
Strings such as "Jan. 1, 2013" fail to parse, due to the presence of the trailing period in the abbreviated month. Since this is a very common way to write out dates, it seems worthwhile to support.

The proposed fix is very simple - just remove all periods from the input text when they're followed by a space.

This does not break any of the current tests, and looking through the current pattern list I don't think it would break any functionality. Due to the requirement of a space after the period, formats such as "d.m.y" are preserved.

Any thoughts?